Move the parts of a big source file to the relevant places.
